Move request details card to own folder
